Refactor utils and tests files. Only moving files around.
